feat(manager-react-components): split create page into steps
ref: DTCORE-2772
